Extract TagColor interface.
